ref(widget-builder-new-experience): Remove prototype files
