--- static/app/views/dashboardsV2/widgetBuilder/eventWidget/queries.tsx
@@ -1,164 +1,0 @@
-import * as React from 'react';
-import styled from '@emotion/styled';
-
-import Button from 'sentry/components/button';
-import SearchBar from 'sentry/components/events/searchBar';
-import Input from 'sentry/components/forms/controls/input';
-import Field from 'sentry/components/forms/field';
-import {IconAdd, IconDelete} from 'sentry/icons';
-import {t} from 'sentry/locale';
-import space from 'sentry/styles/space';
-import {Organization} from 'sentry/types';
-
-import {WidgetQuery} from '../../types';
-import {DisplayType} from '../utils';
-
-type Props = {
-  displayType: DisplayType;
-  onAddQuery: () => void;
-  onChangeQuery: (queryIndex: number, queries: WidgetQuery) => void;
-  onRemoveQuery: (index: number) => void;
-  organization: Organization;
-  queries: WidgetQuery[];
-  selectedProjectIds: number[];
-  errors?: Array<Record<string, any>>;
-};
-
-function Queries({
-  queries,
-  selectedProjectIds,
-  organization,
-  displayType,
-  onRemoveQuery,
-  onAddQuery,
-  onChangeQuery,
-  errors,
-}: Props) {
-  function handleFieldChange(queryIndex: number, field: keyof WidgetQuery) {
-    const widgetQuery = queries[queryIndex];
-    return function handleChange(value: string) {
-      const newQuery = {...widgetQuery, [field]: value};
-      onChangeQuery(queryIndex, newQuery);
-    };
-  }
-
-  function canAddNewQuery() {
-    const rightDisplayType = [
-      DisplayType.LINE,
-      DisplayType.AREA,
-      DisplayType.STACKED_AREA,
-      DisplayType.BAR,
-    ].includes(displayType);
-
-    const underQueryLimit = queries.length < 3;
-    return rightDisplayType && underQueryLimit;
-  }
-
-  const hideLegendAlias = [
-    DisplayType.TABLE,
-    DisplayType.WORLD_MAP,
-    DisplayType.BIG_NUMBER,
-  ].includes(displayType);
-
-  return (
-    <div>
-      {queries.map((query, queryIndex) => {
-        const displayDeleteButton = queries.length > 1;
-        const displayLegendAlias = !hideLegendAlias;
-        return (
-          <StyledField
-            key={queryIndex}
-            inline={false}
-            flexibleControlStateSize
-            stacked
-            error={errors?.[queryIndex].conditions}
-          >
-            <Fields
-              displayDeleteButton={displayDeleteButton}
-              displayLegendAlias={displayLegendAlias}
-            >
-              <SearchBar
-                organization={organization}
-                projectIds={selectedProjectIds}
-                query={query.conditions}
-                fields={[]}
-                onSearch={handleFieldChange(queryIndex, 'conditions')}
-                onBlur={handleFieldChange(queryIndex, 'conditions')}
-                useFormWrapper={false}
-              />
-              {displayLegendAlias && (
-                <Input
-                  type="text"
-                  name="name"
-                  required
-                  value={query.name}
-                  placeholder={t('Legend Alias')}
-                  onChange={event =>
-                    handleFieldChange(queryIndex, 'name')(event.target.value)
-                  }
-                />
-              )}
-              {displayDeleteButton && (
-                <Button
-                  size="zero"
-                  borderless
-                  onClick={event => {
-                    event.preventDefault();
-                    onRemoveQuery(queryIndex);
-                  }}
-                  icon={<IconDelete />}
-                  title={t('Remove query')}
-                  aria-label={t('Remove query')}
-                />
-              )}
-            </Fields>
-          </StyledField>
-        );
-      })}
-      {canAddNewQuery() && (
-        <Button
-          size="small"
-          icon={<IconAdd isCircled />}
-          onClick={(event: React.MouseEvent) => {
-            event.preventDefault();
-            onAddQuery();
-          }}
-        >
-          {t('Add Query')}
-        </Button>
-      )}
-    </div>
-  );
-}
-
-export default Queries;
-
-const fieldsColumns = (p: {
-  displayDeleteButton: boolean;
-  displayLegendAlias: boolean;
-}) => {
-  if (!p.displayDeleteButton && !p.displayLegendAlias) {
-    return '1fr';
-  }
-
-  if (!p.displayDeleteButton) {
-    return '1fr 33%';
-  }
-
-  if (!p.displayLegendAlias) {
-    return '1fr max-content';
-  }
-
-  return '1fr 33% max-content';
-};
-
-const Fields = styled('div')<{displayDeleteButton: boolean; displayLegendAlias: boolean}>`
-  display: grid;
-  grid-template-columns: ${fieldsColumns};
-  gap: ${space(1)};
-  align-items: center;
-`;
-
-const StyledField = styled(Field)`
-  padding-right: 0;
-`;
